Reorganised some of the functions. Still looking for the best way to structure for clarity.
