+++ node/risk-app/server/src/js/map-page/scenario-bars.js
@@ -1,0 +1,21 @@
+function scenarioDisplayUpdate (scenarioBar) {
+  const scenariosRadios = document.querySelectorAll(`input[name="scenarios-${scenarioBar}"]`)
+  scenariosRadios.forEach(radio => {
+    if (radio.checked) {
+      const parent = radio.parentNode
+      parent.style.borderBottom = '7px solid rgb(29, 112, 184)'
+      const scenarioHeading = parent.querySelector('.scenario-heading')
+      scenarioHeading.style.textDecoration = 'none'
+    } else {
+      const parent = radio.parentNode
+      parent.style.borderBottom = 'none'
+      const scenarioHeading = parent.querySelector('.scenario-heading')
+      scenarioHeading.style.textDecoration = 'underline'
+      scenarioHeading.style.textDecorationThickness = '2px'
+    }
+  })
+}
+
+module.exports = {
+  scenarioDisplayUpdate
+}
